Fix names mapping for real
There was an ongoing ambiguity between variables defined by their
scope id vs their scope depth.

This showed up as odd errors during decompilation.

+++ crates/common/src/program/mod.rs
@@ -64,5 +64,5 @@ pub struct Decl {
     pub constant: bool,
     /// The scope id of the variable.
     /// This is used to determine the scope of the variable when binding (or rebinding at decompile)
-    pub scope_id: usize,
+    pub scope_id: u16,
 }
